refactored gilded_rose.rb file
refactoring allows for reduced conditional statements and booleans. This will allow for reduced complexity when adding additional items to inventory, such as the new conjured items. I was not allowed to touch the item class, otherwise I would have reworked this further to allow for items to inherit from a base item class.